+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,8 +1,8 @@
 package com.techcourse.dao;
 
-import com.interface21.jdbc.ObjectMapper;
-import com.interface21.jdbc.PreparedStatementSetter;
 import com.interface21.jdbc.core.JdbcTemplate;
+import com.interface21.jdbc.core.ObjectMapper;
+import com.interface21.jdbc.core.PreparedStatementSetter;
 import com.techcourse.domain.User;
 import java.util.List;
 import javax.sql.DataSource;
@@ -54,7 +54,8 @@ public class UserDao {
 
     public List<User> findAll() {
         final var sql = "select id, account, password, email from users";
-        return jdbcTemplate.query(objectMapper, sql, pstmt -> {});
+        return jdbcTemplate.query(objectMapper, sql, pstmt -> {
+        });
     }
 
     public User findById(final Long id) {
