+++ jdbc/src/main/java/com/interface21/jdbc/core/ObjectMapper.java
@@ -1,10 +1,10 @@
-package com.interface21.jdbc;
+package com.interface21.jdbc.core;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
 @FunctionalInterface
-public interface ObjectMapper<T>  {
+public interface ObjectMapper<T> {
 
     T mapToObject(ResultSet rs) throws SQLException;
 }
