+++ study/src/test/java/transaction/stage1/jdbc/PreparedStatementSetter.java
@@ -4,5 +4,6 @@ import java.sql.PreparedStatement;
 import java.sql.SQLException;
 
 public interface PreparedStatementSetter {
+
     void setParameters(final PreparedStatement pstmt) throws SQLException;
 }
